Removed the target param from some of the json parser methods.
This method is only used when handling the root structure shape.
It can be safely removed from list & map parsing. This resovles
the issue with callers passing a `nil` value that prevents the
default array or hash for list and map. This resolves all of
the test errors in this branch.